+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ????
 - Notes on how to use the HTTP object and class
 - Provide return types
+- New test for changing the default user agent
 
 ## 0.0.4 - 2023-04-12
 - Github Actions for tests